feat: filtrando pedidos #16
